window and gui manager classes
